Search Versioned framework paths
This isn't ideal to use the potentially wrong platform's shared cache
but realistically for most things this is probably good enough, so we
want to find the right frameworks from the macOS cache which uses the
versioned structure, even if the requested framework on iOS does not.
I'm slightly surprised that the shared cache doesn't build these
symlinks, which I guess would be another option after extraction.